Pull latest C code from https://github.com/facebook/zstd/

This is the code as of https://github.com/facebook/zstd/tree/v1.4.8 (https://github.com/facebook/zstd/commit/97a3da1df009d4dc67251de0c4b1c9d7fe286fc1)

Similar to https://github.com/DataDog/zstd/pull/73#issue-337149664 to pull changes with edits to flatten the directory structure