@Viq111 2021-02-11
Benchmark on a Macbook Prop for mr from http://sun.aei.polsl.pl/~sdeor/index.php?page=silesia:

```
ᐅ benchstat before 1.4.8
name                    old time/op   new time/op   delta
CtxCompression-16        91.7ms ± 6%   90.9ms ± 2%    ~     (p=0.721 n=8+8)
CtxDecompression-16      11.7ms ± 1%   10.7ms ± 2%  -8.41%  (p=0.000 n=9+10)
StreamCompression-16     96.4ms ± 2%   99.6ms ± 2%  +3.36%  (p=0.000 n=8+10)
StreamDecompression-16   14.0ms ± 2%   12.8ms ± 2%  -8.81%  (p=0.000 n=10+10)
Compression-16           91.2ms ± 3%   91.4ms ± 2%    ~     (p=0.529 n=10+10)
Decompression-16         11.8ms ± 2%   10.7ms ± 2%  -8.79%  (p=0.000 n=10+10)

name                    old speed     new speed     delta
CtxCompression-16       109MB/s ± 6%  110MB/s ± 2%    ~     (p=0.721 n=8+8)
CtxDecompression-16     851MB/s ± 1%  929MB/s ± 2%  +9.19%  (p=0.000 n=9+10)
StreamCompression-16    103MB/s ± 2%  100MB/s ± 2%  -3.25%  (p=0.000 n=8+10)
StreamDecompression-16  713MB/s ± 2%  782MB/s ± 2%  +9.65%  (p=0.000 n=10+10)
Compression-16          109MB/s ± 3%  109MB/s ± 2%    ~     (p=0.529 n=10+10)
Decompression-16        846MB/s ± 2%  928MB/s ± 2%  +9.64%  (p=0.000 n=10+10)
```

Decompression seems to be ~9% faster
Compression seems to be ~3% slower